+++ public/parcel/appl/js/controller/start.js
@@ -126,7 +126,7 @@ module.exports = App.controllers.Start ||
             finish: function (options) {
                 // var marked = require("marked");
                 var mdFunction = function (data) {
-                    $(".markdown").append(marked.parse(data));
+                    $(".markdown").append(marked.parse(data, {mangle: false, headerIds: false}));
                 };
                 $.get(options.urlMd, mdFunction, "text");
             }