+++ public/parcel/jasmine/toolstest.js
@@ -9,7 +9,7 @@ module.exports = {
             var tools;
             var beforeValue;
             var afterValue;
-            var spyToolsEvent;
+            var spyOnTools;
             var selectorObject;
 
             beforeAll(function (done) {
@@ -27,10 +27,14 @@ module.exports = {
                         fail("The Tools Page did not load within limited time: " + rejected);
                     }).then(function (resolved) {
                         tools = $("#tools");
-                        beforeValue = tools.find("tbody").find("tr:nth-child(1)").find("td:nth-child(2)").text();
+                        const jqueryTools = tools.find("tbody").find("tr:nth-child(1)").find("td:nth-child(2)");
+                        beforeValue = jqueryTools.text();
+                        spyOnTools =
+                            spyOn(jqueryTools,"text").and.callThrough().and.returnValue(beforeValue);
+
+                        beforeValue = jqueryTools.text();
 
                         selectorObject = $('.jobtype-selector');
-                        spyToolsEvent = spyOnEvent(selectorObject[0], 'change');
                         /*
                          *  The can.Component(jobtype-selector) has a change event - we want to test that.
                          */
@@ -41,8 +45,11 @@ module.exports = {
                         // Using RxJs instead of Promise.
                         const numbers = timer(50, 50);
                         const observable = numbers.subscribe(timer => {
-                            afterValue = tools.find('tbody').find('tr:nth-child(1)').find('td:nth-child(2)').text()
-                            if (afterValue !== beforeValue || timer === 15) {
+                            afterValue = tools.find("tbody").find("tr:nth-child(1)").find("td:nth-child(2)").text();
+                            spyOnTools.and.callThrough().and.returnValue(afterValue);
+                            afterValue = jqueryTools.text();
+                            if (spyOnTools.calls.first().returnValue !==
+                                    spyOnTools.calls.mostRecent().returnValue || timer === 15) {
                                 observable.unsubscribe();
                                 done();
                             }
@@ -51,14 +58,9 @@ module.exports = {
             });
 
             it("setup and change event executed.", function (done) {
-                // jasmine-jquery matchers
-                expect('change').toHaveBeenTriggeredOn(selectorObject[0]);
-                expect(spyToolsEvent).toHaveBeenTriggered();
-
                 expect(tools[0]).toBeInDOM();
                 expect('.disabled').toBeDisabled();
 
-                // expect(selectorObject.focus()).toBeFocused();
                 // Required for Firefox
                 selectorObject[0] = document.activeElement;
                 expect(selectorObject).toBeFocused();
@@ -69,7 +71,11 @@ module.exports = {
 
             it("new page loaded on change.", function (done) {
                 // Verify that new page was loaded.
-                expect(beforeValue).not.toBe(afterValue);
+                expect(spyOnTools.calls.first().returnValue.length > 0).toBe(true);
+                expect(spyOnTools.calls.mostRecent().returnValue.length > 0).toBe(true);
+                expect(spyOnTools).toHaveBeenCalled();
+                expect(spyOnTools.calls.count() > 1).toBe(true);
+                expect(spyOnTools.calls.first().returnValue).not.toBe(spyOnTools.calls.mostRecent().returnValue);
 
                 done();
             });