+++ server/routes/dodex.js
@@ -1,7 +1,7 @@
 /**
  * This is the websocket route. Does all messaging and database calls.
  */
-const database = require("../db/database");
+const database = require("../db/database-obj");
 const orm = database.orm;
 const Router = require("koa-router");
 const Url = require("url").URL;
@@ -74,7 +74,7 @@ function socketServer(server) {
     ws.handle = handle;
 
     await orm.getUser(ws).then(async (record) => {
-      if (typeof record === "undefined" || record === null) {
+      if (typeof record === "undefined" || record === null || record.length === 0) {
         await orm.addUser(ws, user)
           .catch((err) => {
             utils.log("error", err.stack, __filename);
@@ -83,7 +83,8 @@ function socketServer(server) {
             throw err.message;
           });
       }
-      else if (record.get("password") !== id) {
+//      else if (record.get("password") !== id) {
+      else if (record[0].password !== id) {
         ws.send(userErrorMessage);
         throw `ID does not match - ${id} for ${handle}`;
       }
@@ -184,7 +185,8 @@ function undeliveredPrivateMessages(ws, disconnectedUsers, data) {
     return;
   }
   orm.addMessage(ws, data).then(async (message) => {
-    await orm.addUndelivered(ws, disconnectedUsers, message.get("id"));
+//    await orm.addUndelivered(ws, disconnectedUsers, message.get("id"));
+    await orm.addUndelivered(ws, disconnectedUsers, message.id);
   })
     .catch((err) => {
       utils.log("error", err.message, __filename);